afnetns: replace net pointer in sock_common with afnetns if configured
The sockets need to store their current afnetns for bind and connect
operations. Because each afnetns has a dependency on its network
namespace, we can simply remove the skc_net pointers from struct sock
and forward them to the afnetns struct.

We store net references without taking a reference count at several
locations: tcp_metric blocks, request sockets and time wait sockets. We
need to adhere to their lifetimes guarantees as follows:

* tcp metric blocks will still only be stored per net namespace and
  don't take a reference to the afnetns which they were created in. This
  fits with afnetns design, because the demultiplexing can be done on
  the destination IP addresses and thus all entries are unique.

* timewait sockets are flushed during teardown of the network namespace
  and we must adhere to the same lifetime guarantees. During shutdown of
  an afnetns we call to tcp and flush the appropriate sockets. Normally
  timewait state is uniquly identified by quad tuple, so we strictly
  would not need to care about afnetns here, but for the sake of reusing
  common infrastructure and not creating more special cases, we reuse
  the current logic.

* inet request socket's lifetime is bound by their appropriate
  sk_listener, which takes a reference on the afnetns. If the sk_listener
  is shut down, we flush all request sockets, thus no further handling
  is needed here and everything works out fine already.

The reason for the ifdef's is that we want to make sure to not change
any logic later on in case afnetns is in anyway flawed. I plan to remove
the ifdef's later on after afnetns got more testing in field. Because of
its minimal overhead, we might make it default in case network namespaces
are enabled.

Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>