afnetns: store afnetns in in_ifaddr and inet6_ifaddr
Accept afnetns parameters via the netlink interface while adding ip(v6)
addresses. If no parameter is given, we use the current process'
afnetns context (for ioctl as well as netlink interface).

For IPv6 autoconfiguration we default to the default afnetns for the
namespace.

The API regarding setting and querying the afnetns is asymetric. We use
file descriptors to afnetns objects to set up the IP address but the
result of querying is an inode number. The reason for that is that we
want to eliminate races during set up. Because we can't exclude races
during querying and we can't modify the caller's file table, we report
back an inode and let iproute or other tools figure out which afnetns
the ip address is attached to. iproute will do that automatically for you.

Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>